requirements.txt
streamlit
opencv-python-headless
numpy
tensorflow
pillow
matplotlib
